Revert "Upgrade TF version from 2.7 to 2.11 (#749)"
This reverts commit 20cfde5eb2f6188d35bab1375c98572ae4571b20.

+++ tensorflow_quantum/core/ops/cirq_ops_test.py
@@ -457,7 +457,7 @@ class CirqSamplesTest(tf.test.TestCase, parameterized.TestCase):
                             'tfq':
                                 np.array([[1] * len(program.all_qubits())] *
                                          repetitions,
-                                         dtype=int),
+                                         dtype=np.int32),
                         }) for param in cirq.to_resolvers(params)
                 ]
